refactor(dashboard): split quick-trade layer out of live-quotes view-model
The retry-gating fix pushed live-quotes-screen.view-model.ts past the
no-new-god-file ratchet cap (2637 lines). Extract the quick-trade ticket
layer — form state, validation, submission lifecycle, staged-review view
models, and their helpers — into live-quotes-screen.quick-trade.ts. The
view-model module re-exports it, so all existing import paths are unchanged
and behavior is identical.
